Refactor legend configuration: Move legend-related calculations to legendConfiguration.ts
- Moved functions for calculating bottom legend height and right legend width from convertDataIntoUnitValue.ts to legendConfiguration.ts for better organization.
- Updated imports in convertPromQLData.ts and convertSQLData.ts to reflect the new location of legend calculation functions.
- Removed unused imports and interfaces in legendConfiguration.ts to streamline the code.
- Added utility functions for calculating legend dimensions and chart dimensions based on legend configuration.
