+++ web/src/components/dashboards/addPanel/ConfigPanel.vue
@@ -244,7 +244,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
     </div>
 
     <q-toggle
-      v-if="showLegendsConditions"
+      v-if="shouldShowLegendsToggle(dashboardPanelData)"
       v-model="dashboardPanelData.data.config.show_legends"
       :label="t('dashboard.showLegendsLabel')"
       data-test="dashboard-config-show-legend"
@@ -281,7 +281,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
     <div class="o2-input">
       <q-select
-        v-if="showLegendPositionConditions"
+        v-if="shouldShowLegendPosition(dashboardPanelData)"
         outlined
         v-model="dashboardPanelData.data.config.legends_position"
         :options="legendsPositionOptions"
@@ -300,7 +300,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
       <div class="space"></div>
 
       <q-select
-        v-if="showLegendTypeConditions"
+        v-if="shouldShowLegendType(dashboardPanelData)"
         outlined
         v-model="dashboardPanelData.data.config.legends_type"
         :options="legendTypeOptions"
@@ -321,7 +321,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
       <div class="input-container">
         <!-- Legend Width Configuration (for right position) -->
         <q-input
-          v-if="showLegendWidthConditions"
+          v-if="shouldShowLegendWidth(dashboardPanelData)"
           v-model.number="legendWidthValue"
           :label="t('common.legendWidth')"
           color="input-border"
@@ -339,7 +339,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
         <!-- Legend Height Configuration (for auto/bottom position) -->
         <q-input
-          v-if="showLegendHeightConditions"
+          v-if="shouldShowLegendHeight(dashboardPanelData)"
           v-model.number="legendHeightValue"
           label="Legend Height"
           color="input-border"
@@ -358,7 +358,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
         <!-- Unit container for Legend Width (right position) -->
         <div
           class="unit-container"
-          v-if="showLegendWidthUnitContainerConditions"
+          v-if="shouldShowLegendWidthUnitContainer(dashboardPanelData)"
         >
           <button
             @click="setUnit('px')"
@@ -396,7 +396,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
         <!-- Unit container for Legend Height (auto/bottom position) -->
         <div
           class="unit-container"
-          v-if="showLegendHeightUnitContainerConditions"
+          v-if="shouldShowLegendHeightUnitContainer(dashboardPanelData)"
         >
           <button
             @click="setHeightUnit('px')"
@@ -435,16 +435,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
       <div class="space"></div>
 
       <q-select
-        v-if="
-          (dashboardPanelData.data.type === 'pie' ||
-            dashboardPanelData.data.type === 'donut') &&
-          dashboardPanelData.data.config.show_legends &&
-          dashboardPanelData.data.config.legends_position === 'right' &&
-          (dashboardPanelData.data.config.legends_type === 'plain' ||
-            dashboardPanelData.data.config.legends_type === 'scroll' ||
-            dashboardPanelData.data.config.legends_type === null) &&
-          !isTrellisEnabled
-        "
+        v-if="shouldApplyChartAlign(dashboardPanelData)"
         outlined
         v-model="dashboardPanelData.data.config.chart_align"
         :options="chartAlignOptions"
@@ -454,13 +445,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
         stack-label
         emit-value
         :display-value="`${
-          dashboardPanelData.data.config.chart_align === null
-            ? 'Auto'
-            : dashboardPanelData.data.config.chart_align === 'left'
-              ? 'Left'
-              : dashboardPanelData.data.config.chart_align === 'center'
-                ? 'Center'
-                : 'Auto'
+          dashboardPanelData.data.config.chart_align ?? 'Auto'
         }`"
         data-test="dashboard-config-chart-align"
       >
@@ -1351,17 +1336,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
       <div class="space"></div>
 
       <q-toggle
-        v-if="
-          dashboardPanelData.data.type != 'table' &&
-          dashboardPanelData.data.type != 'heatmap' &&
-          dashboardPanelData.data.type != 'metric' &&
-          dashboardPanelData.data.type != 'gauge' &&
-          dashboardPanelData.data.type != 'geomap' &&
-          dashboardPanelData.data.type != 'pie' &&
-          dashboardPanelData.data.type != 'donut' &&
-          dashboardPanelData.data.type != 'sankey' &&
-          dashboardPanelData.data.type != 'maps'
-        "
+        v-if="shouldShowGridlines(dashboardPanelData)"
         v-model="dashboardPanelData.data.config.show_gridlines"
         label="Show Gridlines"
         data-test="dashboard-config-show-gridlines"
@@ -1652,14 +1627,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 <script lang="ts">
 import useDashboardPanelData from "@/composables/useDashboardPanel";
-import {
-  computed,
-  defineComponent,
-  inject,
-  onBeforeMount,
-  onMounted,
-  ref,
-} from "vue";
+import { computed, defineComponent, inject, onBeforeMount } from "vue";
 import { useI18n } from "vue-i18n";
 import Drilldown from "./Drilldown.vue";
 import ValueMapping from "./ValueMapping.vue";
@@ -1687,6 +1655,8 @@ import {
   shouldShowLegendHeight,
   shouldShowLegendWidthUnitContainer,
   shouldShowLegendHeightUnitContainer,
+  shouldApplyChartAlign,
+  shouldShowGridlines,
 } from "@/utils/dashboard/configUtils";
 
 export default defineComponent({
@@ -1716,6 +1686,7 @@ export default defineComponent({
     const { dashboardPanelData, promqlMode } = useDashboardPanelData(
       dashboardPanelDataPageKey,
     );
+
     const { t } = useI18n();
     const store = useStore();
 
@@ -2307,45 +2278,6 @@ export default defineComponent({
       );
     });
 
-    const isTrellisEnabled = computed(() => {
-      return dashboardPanelData.data.config.trellis?.layout !== null;
-    });
-
-    // Computed properties for legend conditions
-    const showLegendsConditions = computed(() =>
-      shouldShowLegendsToggle(dashboardPanelData, isTrellisEnabled.value),
-    );
-
-    const showLegendPositionConditions = computed(() =>
-      shouldShowLegendPosition(dashboardPanelData, isTrellisEnabled.value),
-    );
-
-    const showLegendTypeConditions = computed(() =>
-      shouldShowLegendType(dashboardPanelData, isTrellisEnabled.value),
-    );
-
-    const showLegendWidthConditions = computed(() =>
-      shouldShowLegendWidth(dashboardPanelData, isTrellisEnabled.value),
-    );
-
-    const showLegendHeightConditions = computed(() =>
-      shouldShowLegendHeight(dashboardPanelData, isTrellisEnabled.value),
-    );
-
-    const showLegendWidthUnitContainerConditions = computed(() =>
-      shouldShowLegendWidthUnitContainer(
-        dashboardPanelData,
-        isTrellisEnabled.value,
-      ),
-    );
-
-    const showLegendHeightUnitContainerConditions = computed(() =>
-      shouldShowLegendHeightUnitContainer(
-        dashboardPanelData,
-        isTrellisEnabled.value,
-      ),
-    );
-
     // Clear legend width when switching away from plain type or when position is not right
     watchEffect(() => {
       if (
@@ -2403,15 +2335,16 @@ export default defineComponent({
       showTrellisConfig,
       isBreakdownFieldEmpty,
       hasTimeShifts,
-      isTrellisEnabled,
       dashboardPanelDataPageKey,
-      showLegendsConditions,
-      showLegendPositionConditions,
-      showLegendTypeConditions,
-      showLegendWidthConditions,
-      showLegendHeightConditions,
-      showLegendWidthUnitContainerConditions,
-      showLegendHeightUnitContainerConditions,
+      shouldShowLegendsToggle,
+      shouldShowLegendPosition,
+      shouldShowLegendType,
+      shouldShowLegendWidth,
+      shouldShowLegendHeight,
+      shouldShowLegendWidthUnitContainer,
+      shouldShowLegendHeightUnitContainer,
+      shouldApplyChartAlign,
+      shouldShowGridlines,
     };
   },
 });
