+++ web/src/utils/dashboard/configUtils.ts
@@ -2,27 +2,49 @@
  * Utility functions for dashboard panel configuration conditions
  */
 
-interface DashboardPanelData {
-  data: {
-    type: string;
-    config: {
-      show_legends?: boolean;
-      legends_position?: string | null;
-      legends_type?: string | null;
-    };
-  };
-}
+/**
+ * Check that chart align option should be applied
+ * @param dashboardPanelData - The dashboard panel data
+ * @returns boolean indicating if chart align option should be applied
+ */
+export const shouldApplyChartAlign = (dashboardPanelData: any): boolean => {
+  return (
+    (dashboardPanelData.data.type === "pie" ||
+      dashboardPanelData.data.type === "donut") &&
+    dashboardPanelData.data.config.show_legends &&
+    dashboardPanelData.data.config.legends_position === "right" &&
+    (dashboardPanelData.data.config.legends_type === "plain" ||
+      dashboardPanelData.data.config.legends_type === "scroll" ||
+      dashboardPanelData.data.config.legends_type === null) &&
+    dashboardPanelData.data.config.trellis?.layout === null
+  );
+};
+
+/**
+ * Check that show Gridlines option should be applied
+ * @param dashboardPanelData - The dashboard panel data
+ * @returns boolean indicating if show Gridlines option should be applied
+ */
+export const shouldShowGridlines = (dashboardPanelData: any): boolean => {
+  return (
+    dashboardPanelData.data.type != "table" &&
+    dashboardPanelData.data.type != "heatmap" &&
+    dashboardPanelData.data.type != "metric" &&
+    dashboardPanelData.data.type != "gauge" &&
+    dashboardPanelData.data.type != "geomap" &&
+    dashboardPanelData.data.type != "pie" &&
+    dashboardPanelData.data.type != "donut" &&
+    dashboardPanelData.data.type != "sankey" &&
+    dashboardPanelData.data.type != "maps"
+  );
+};
 
 /**
  * Check if show legends toggle should be displayed
- * @param dashboardPanelData - The dashboard panel data object
- * @param isTrellisEnabled - Whether trellis is enabled
+ * @param dashboardPanelData - The dashboard panel data
  * @returns boolean indicating if show legends should be displayed
  */
-export function shouldShowLegendsToggle(
-  dashboardPanelData: DashboardPanelData,
-  isTrellisEnabled: boolean,
-): boolean {
+export const shouldShowLegendsToggle = (dashboardPanelData: any): boolean => {
   return (
     dashboardPanelData.data.type !== "table" &&
     dashboardPanelData.data.type !== "heatmap" &&
@@ -31,20 +53,16 @@ export function shouldShowLegendsToggle(
     dashboardPanelData.data.type !== "geomap" &&
     dashboardPanelData.data.type !== "sankey" &&
     dashboardPanelData.data.type !== "maps" &&
-    !isTrellisEnabled
+    dashboardPanelData.data.config.trellis?.layout === null
   );
-}
+};
 
 /**
  * Check if legend position selector should be displayed
- * @param dashboardPanelData - The dashboard panel data object
- * @param isTrellisEnabled - Whether trellis is enabled
+ * @param dashboardPanelData - The dashboard panel data
  * @returns boolean indicating if legend position should be displayed
  */
-export function shouldShowLegendPosition(
-  dashboardPanelData: DashboardPanelData,
-  isTrellisEnabled: boolean,
-): boolean {
+export const shouldShowLegendPosition = (dashboardPanelData: any): boolean => {
   return (
     dashboardPanelData.data.type !== "table" &&
     dashboardPanelData.data.type !== "heatmap" &&
@@ -54,20 +72,16 @@ export function shouldShowLegendPosition(
     dashboardPanelData.data.config.show_legends &&
     dashboardPanelData.data.type !== "sankey" &&
     dashboardPanelData.data.type !== "maps" &&
-    !isTrellisEnabled
+    dashboardPanelData.data.config.trellis?.layout === null
   );
-}
+};
 
 /**
  * Check if legend type selector should be displayed
- * @param dashboardPanelData - The dashboard panel data object
- * @param isTrellisEnabled - Whether trellis is enabled
+ * @param dashboardPanelData - The dashboard panel data
  * @returns boolean indicating if legend type should be displayed
  */
-export function shouldShowLegendType(
-  dashboardPanelData: DashboardPanelData,
-  isTrellisEnabled: boolean,
-): boolean {
+export const shouldShowLegendType = (dashboardPanelData: any): boolean => {
   return (
     dashboardPanelData.data.type !== "table" &&
     dashboardPanelData.data.type !== "heatmap" &&
@@ -77,20 +91,15 @@ export function shouldShowLegendType(
     dashboardPanelData.data.config.show_legends &&
     dashboardPanelData.data.type !== "sankey" &&
     dashboardPanelData.data.type !== "maps" &&
-    !isTrellisEnabled
+    dashboardPanelData.data.config.trellis?.layout === null
   );
-}
-
+};
 /**
  * Check if legend width configuration should be displayed
- * @param dashboardPanelData - The dashboard panel data object
- * @param isTrellisEnabled - Whether trellis is enabled
+ * @param dashboardPanelData - The dashboard panel data
  * @returns boolean indicating if legend width should be displayed
  */
-export function shouldShowLegendWidth(
-  dashboardPanelData: DashboardPanelData,
-  isTrellisEnabled: boolean,
-): boolean {
+export const shouldShowLegendWidth = (dashboardPanelData: any): boolean => {
   return (
     dashboardPanelData.data.type !== "table" &&
     dashboardPanelData.data.type !== "heatmap" &&
@@ -101,20 +110,16 @@ export function shouldShowLegendWidth(
     dashboardPanelData.data.config.legends_position === "right" &&
     dashboardPanelData.data.config.legends_type === "plain" &&
     dashboardPanelData.data.type !== "sankey" &&
-    !isTrellisEnabled
+    dashboardPanelData.data.config.trellis?.layout === null
   );
-}
+};
 
 /**
  * Check if legend height configuration should be displayed
- * @param dashboardPanelData - The dashboard panel data object
- * @param isTrellisEnabled - Whether trellis is enabled
+ * @param dashboardPanelData - The dashboard panel data
  * @returns boolean indicating if legend height should be displayed
  */
-export function shouldShowLegendHeight(
-  dashboardPanelData: DashboardPanelData,
-  isTrellisEnabled: boolean,
-): boolean {
+export const shouldShowLegendHeight = (dashboardPanelData: any): boolean => {
   return (
     dashboardPanelData.data.type !== "table" &&
     dashboardPanelData.data.type !== "heatmap" &&
@@ -126,32 +131,28 @@ export function shouldShowLegendHeight(
       dashboardPanelData.data.config.legends_position === "bottom") &&
     dashboardPanelData.data.config.legends_type === "plain" &&
     dashboardPanelData.data.type !== "sankey" &&
-    !isTrellisEnabled
+    dashboardPanelData.data.config.trellis?.layout === null
   );
-}
+};
 
 /**
  * Check if legend width unit container should be displayed
- * @param dashboardPanelData - The dashboard panel data object
- * @param isTrellisEnabled - Whether trellis is enabled
+ * @param dashboardPanelData - The dashboard panel data
  * @returns boolean indicating if legend width unit container should be displayed
  */
-export function shouldShowLegendWidthUnitContainer(
-  dashboardPanelData: DashboardPanelData,
-  isTrellisEnabled: boolean,
-): boolean {
-  return shouldShowLegendWidth(dashboardPanelData, isTrellisEnabled);
-}
+export const shouldShowLegendWidthUnitContainer = (
+  dashboardPanelData: any,
+): boolean => {
+  return shouldShowLegendWidth(dashboardPanelData);
+};
 
 /**
  * Check if legend height unit container should be displayed
- * @param dashboardPanelData - The dashboard panel data object
- * @param isTrellisEnabled - Whether trellis is enabled
+ * @param dashboardPanelData - The dashboard panel data
  * @returns boolean indicating if legend height unit container should be displayed
  */
-export function shouldShowLegendHeightUnitContainer(
-  dashboardPanelData: DashboardPanelData,
-  isTrellisEnabled: boolean,
-): boolean {
-  return shouldShowLegendHeight(dashboardPanelData, isTrellisEnabled);
-}
+export const shouldShowLegendHeightUnitContainer = (
+  dashboardPanelData: any,
+): boolean => {
+  return shouldShowLegendHeight(dashboardPanelData);
+};
