+++ web/src/utils/dashboard/legendConfiguration.ts
@@ -13,21 +13,9 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-import {
-  calculateBottomLegendHeight,
-  calculateRightLegendWidth,
-  calculateChartDimensions,
-  generateChartAlignmentProperties,
-  shouldApplyChartAlignment,
-  calculateWidthText,
-} from "./convertDataIntoUnitValue";
+import { calculateWidthText } from "./convertDataIntoUnitValue";
 
 // TypeScript interfaces for better type safety
-interface LegendDimensions {
-  width: number;
-  height: number;
-}
-
 interface ChartDimensions {
   chartWidth: number;
   chartHeight: number;
@@ -63,12 +51,6 @@ interface LegendOptions {
   formatter?: (name: any) => string;
 }
 
-interface ChartOptions {
-  legend: LegendOptions;
-  grid: any;
-  series: any[];
-}
-
 // Constants for better maintainability
 const LEGEND_CONSTANTS = {
   DEFAULT_WIDTH: 800,
@@ -83,13 +65,6 @@ const LEGEND_CONSTANTS = {
   MAX_SCROLL_WIDTH: 150,
   MIN_SCROLL_RESERVED: 120,
 } as const;
-
-const CHART_TYPE_GROUPS = {
-  PIE_DONUT: ["pie", "donut"] as const,
-  EXCLUDED_STANDARD: ["gauge", "metric", "pie", "donut"] as const,
-} as const;
-
-type ExcludedChartType = (typeof CHART_TYPE_GROUPS.EXCLUDED_STANDARD)[number];
 
 /**
  * Utility function to get chart dimensions safely
@@ -111,101 +86,6 @@ const shouldShowLegends = (panelSchema: PanelSchema): boolean =>
     panelSchema.config?.trellis?.layout === undefined);
 
 /**
- * Utility function to check if chart type is excluded from standard legend processing
- */
-const isExcludedChartType = (chartType: string): boolean =>
-  CHART_TYPE_GROUPS.EXCLUDED_STANDARD.includes(chartType as any);
-
-/**
- * Utility function to check if explicit legend dimension is set
- */
-const hasExplicitLegendWidth = (panelSchema: PanelSchema): boolean =>
-  Boolean(
-    panelSchema.config?.legend_width &&
-      !isNaN(parseFloat(panelSchema.config.legend_width.value.toString())),
-  );
-
-const hasExplicitLegendHeight = (panelSchema: PanelSchema): boolean =>
-  Boolean(
-    panelSchema.config?.legend_height &&
-      !isNaN(parseFloat(panelSchema.config.legend_height.value.toString())),
-  );
-
-/**
- * Utility function to calculate legend width from config
- */
-const calculateLegendWidthFromConfig = (
-  panelSchema: PanelSchema,
-  chartWidth: number,
-): number => {
-  const widthConfig = panelSchema.config?.legend_width;
-  if (!widthConfig) return 0;
-
-  return widthConfig.unit === "%"
-    ? chartWidth * (widthConfig.value / 100)
-    : widthConfig.value;
-};
-
-/**
- * Utility function to calculate legend height from config
- */
-const calculateLegendHeightFromConfig = (
-  panelSchema: PanelSchema,
-  chartHeight: number,
-): number => {
-  const heightConfig = panelSchema.config?.legend_height;
-  if (!heightConfig) return 0;
-
-  return heightConfig.unit === "%"
-    ? chartHeight * (heightConfig.value / 100)
-    : heightConfig.value;
-};
-
-/**
- * Shared condition checker for right legend positioning
- */
-const shouldApplyRightLegendPositioning = (
-  panelSchema: PanelSchema,
-  isPlainOrExplicit: boolean,
-): boolean =>
-  shouldShowLegends(panelSchema) &&
-  !CHART_TYPE_GROUPS.EXCLUDED_STANDARD.includes(panelSchema.type as any) &&
-  panelSchema.config?.legends_position === "right" &&
-  isPlainOrExplicit;
-
-/**
- * Calculates reserved width for scroll legends based on container size
- */
-const calculateScrollLegendReservedWidth = (chartWidth: number): number => {
-  let reservedWidthPercentage = 0.25; // Default 25%
-  if (chartWidth < 400) {
-    reservedWidthPercentage = 0.3; // Use 30% for very small containers
-  } else if (chartWidth < 600) {
-    reservedWidthPercentage = 0.4; // Use 40% for small containers
-  }
-
-  const minScrollLegendWidth = Math.min(
-    chartWidth * reservedWidthPercentage,
-    LEGEND_CONSTANTS.MAX_SCROLL_WIDTH,
-  );
-  return Math.max(minScrollLegendWidth, LEGEND_CONSTANTS.MIN_SCROLL_WIDTH);
-};
-
-/**
- * Applies legend positioning to the right side of the chart
- */
-const applyRightSidePositioning = (
-  options: ChartOptions,
-  legendWidth: number,
-  containerWidth: number,
-): void => {
-  options.grid.right = legendWidth;
-  const legendLeftPx = Math.max(containerWidth - legendWidth, 0);
-  options.legend.left = legendLeftPx;
-  options.legend.right = 0;
-};
-
-/**
  * Converts the legend position string to the format expected by ECharts.
  *
  * @param {string} legendPosition - The desired position of the legend. Possible values are "bottom" and "right".
@@ -220,6 +100,457 @@ export const getLegendPosition = (legendPosition?: string): string => {
     default:
       return "horizontal";
   }
+};
+
+/**
+ * Converts rem units to pixels based on the root font size (typically 16px)
+ * @param {number} rem - The rem value to convert
+ * @param {number} rootFontSize - The root font size in pixels (default: 16)
+ * @returns {number} The value in pixels
+ */
+const remToPx = (rem: number, rootFontSize: number = 16): number => {
+  return rem * rootFontSize;
+};
+
+/**
+ * Calculates and configures the height needed for legends when positioned at the bottom
+ * @param {number} legendCount - Number of legend items
+ * @param {number} chartWidth - Width of the chart container
+ * @param {any[]} seriesData - Series data to get actual legend names
+ * @param {number} maxHeight - Maximum available height for chart container (optional)
+ * @param {any} legendConfig - Legend configuration object to modify (optional)
+ * @param {any} gridConfig - Grid configuration object to modify (optional)
+ * @param {number} chartHeight - Height of the chart container (required if configuring position)
+ * @returns {number} Height in pixels to reserve at the bottom
+ */
+export const calculateBottomLegendHeight = (
+  legendCount: number,
+  chartWidth: number,
+  seriesData: any[] = [],
+  maxHeight?: number,
+  legendConfig?: any,
+  gridConfig?: any,
+  chartHeight?: number,
+): number => {
+  if (legendCount === 0) return 0;
+
+  // Constants for legend sizing in rem units (converted to pixels)
+  const LEGEND_ITEM_HEIGHT = remToPx(1.25); // 1.25rem = 20px (Height per legend item row)
+  const LEGEND_PADDING = remToPx(1.5); // 1.5rem = 24px (Top and bottom padding)
+  const LEGEND_ICON_WIDTH = remToPx(0.875); // 0.875rem = 14px (Width of legend icon/symbol)
+  const LEGEND_ICON_MARGIN = remToPx(0.5); // 0.5rem = 8px (Margin between icon and text)
+  const LEGEND_ITEM_MARGIN = remToPx(1.25); // 1.25rem = 20px (Horizontal margin between legend items)
+  const MIN_TEXT_WIDTH = remToPx(3.125); // 3.125rem = 50px (Minimum text width per legend item)
+  const MAX_TEXT_WIDTH = remToPx(9.375); // 9.375rem = 150px (Maximum text width per legend item)
+
+  // Calculate average text width based on actual series names
+  let avgTextWidth = MIN_TEXT_WIDTH;
+  if (seriesData.length > 0) {
+    const totalTextWidth = seriesData.reduce((sum, series) => {
+      const name = (series.name || series.seriesName || "").toString();
+      // Use calculateWidthText for accurate font-based measurements
+      const textWidth = calculateWidthText(name, "12px");
+      return sum + textWidth;
+    }, 0);
+    const avgTextLength = totalTextWidth / seriesData.length;
+    // Use actual calculated width, constrained by min/max
+    avgTextWidth = Math.min(
+      Math.max(avgTextLength, MIN_TEXT_WIDTH),
+      MAX_TEXT_WIDTH,
+    );
+  }
+
+  // Calculate total width needed per legend item
+  const itemWidth =
+    LEGEND_ICON_WIDTH + LEGEND_ICON_MARGIN + avgTextWidth + LEGEND_ITEM_MARGIN;
+
+  // Calculate how many items can fit per row
+  const availableWidth = chartWidth - LEGEND_PADDING * 2;
+  const itemsPerRow = Math.max(1, Math.floor(availableWidth / itemWidth));
+
+  // Calculate number of rows needed
+  const numRows = Math.ceil(legendCount / itemsPerRow);
+
+  // Calculate total height
+  const totalHeight = numRows * LEGEND_ITEM_HEIGHT + LEGEND_PADDING;
+
+  // Apply 50% maximum height constraint if maxHeight is provided
+  // This ensures 50% of space is reserved for the chart and 50% maximum for legends
+  let finalHeight = Math.max(totalHeight, remToPx(3.125)); // Minimum height of 3.125rem = 50px
+
+  if (maxHeight && maxHeight > 0) {
+    const maxAllowedHeight = maxHeight * 0.5; // 50% maximum for legends, 50% for chart
+    finalHeight = Math.min(finalHeight, maxAllowedHeight);
+  }
+
+  // If configuration objects are provided, also configure positioning
+  if (legendConfig && gridConfig && chartHeight) {
+    // Set grid bottom margin
+    gridConfig.bottom = finalHeight;
+
+    // Position legend at exact location to prevent overflow to top
+    const legendTopPosition = chartHeight - finalHeight + 10; // 10px padding from bottom
+    legendConfig.top = legendTopPosition;
+    legendConfig.height = finalHeight - 20; // Constrain height within allocated space
+  }
+
+  return finalHeight;
+};
+
+/**
+ * Calculates the width needed for legends when positioned on the right
+ * Keeps chart height unchanged and ensures legends do not overlap the plot
+ * @param {number} legendCount - Number of legend items
+ * @param {number} chartWidth - Width of the chart container
+ * @param {number} chartHeight - Height of the chart container
+ * @param {any[]} seriesData - Series data to get actual legend names
+ * @param {boolean} isScrollable - Whether the legend is scrollable
+ * @returns {number} Width in pixels to reserve on the right
+ */
+export const calculateRightLegendWidth = (
+  legendCount: number,
+  chartWidth: number,
+  chartHeight: number,
+  seriesData: any[] = [],
+  isScrollable: boolean = false,
+): number => {
+  if (legendCount === 0) return 0;
+
+  // Sizing constants in rem units (converted to pixels)
+  const LEGEND_ROW_HEIGHT = remToPx(1.125); // 1.125rem = 18px (per line height)
+  const LEGEND_ICON_WIDTH = remToPx(0.875); // 0.875rem = 14px
+  const LEGEND_ICON_MARGIN = remToPx(0.625); // 0.625rem = 10px
+  const HORIZONTAL_PADDING = remToPx(1.5); // 1.5rem = 24px (left+right padding for legend area)
+  const VERTICAL_PADDING = remToPx(1); // 1rem = 16px
+  const MIN_TEXT_WIDTH = remToPx(5); // 5rem = 80px
+  const MAX_TEXT_WIDTH = remToPx(11.25); // 11.25rem = 180px (allow wider names)
+
+  // Longest text width to avoid truncation
+  let longestTextWidth = MIN_TEXT_WIDTH;
+  if (seriesData.length > 0) {
+    longestTextWidth = seriesData.reduce((max, series) => {
+      const name = (series.name || series.seriesName || "").toString();
+      const width = calculateWidthText(name);
+      return Math.max(max, width);
+    }, MIN_TEXT_WIDTH);
+    longestTextWidth = Math.min(
+      Math.max(longestTextWidth, MIN_TEXT_WIDTH),
+      MAX_TEXT_WIDTH,
+    );
+  }
+
+  // Width per legend column (icon + gap + text)
+  const perColumnWidth =
+    LEGEND_ICON_WIDTH + LEGEND_ICON_MARGIN + longestTextWidth;
+
+  // Determine rows that fit vertically and required columns
+  const availableHeight = Math.max(
+    chartHeight - VERTICAL_PADDING * 2,
+    LEGEND_ROW_HEIGHT,
+  );
+  const rows = Math.max(1, Math.floor(availableHeight / LEGEND_ROW_HEIGHT));
+  const cols = isScrollable ? 1 : Math.ceil(legendCount / rows);
+
+  // Total width reserved for legends
+  const totalLegendWidth = cols * perColumnWidth + HORIZONTAL_PADDING * 2;
+
+  // Apply constraint: max 50% of chart width to ensure chart remains visible
+  const maxAllowedWidth = chartWidth * 0.5;
+  return Math.min(totalLegendWidth, maxAllowedWidth);
+};
+
+/**
+ * Calculates available space dimensions for charts based on legend configuration
+ * @param {any} panelSchema - Panel configuration schema
+ * @param {number} chartWidth - Original chart width
+ * @param {number} chartHeight - Original chart height
+ * @param {any[]} seriesData - Series data for legend calculations
+ * @returns {object} Available dimensions and legend space info
+ */
+export const calculateChartDimensions = (
+  panelSchema: any,
+  chartWidth: number,
+  chartHeight: number,
+  seriesData: any[] = [],
+) => {
+  const config = panelSchema.config || {};
+  const legendPosition = config.legends_position;
+  const showLegends = config.show_legends;
+  const legendCount = seriesData?.length || 0;
+
+  let availableWidth = chartWidth;
+  let availableHeight = chartHeight;
+  let legendWidth = 0;
+  let legendHeight = 0;
+
+  // Early return if no legends to display
+  if (!showLegends || legendCount === 0) {
+    return {
+      availableWidth,
+      availableHeight,
+      legendWidth,
+      legendHeight,
+      hasLegends: false,
+    };
+  }
+
+  const hasExplicitWidth =
+    config.legend_width && !isNaN(parseFloat(config.legend_width.value));
+  const hasExplicitHeight =
+    config.legend_height && !isNaN(parseFloat(config.legend_height.value));
+
+  // Calculate legend dimensions based on position and type
+  if (legendPosition === "right") {
+    legendWidth = calculateLegendWidth(
+      panelSchema,
+      chartWidth,
+      chartHeight,
+      seriesData,
+      hasExplicitWidth,
+    );
+    availableWidth = chartWidth - legendWidth;
+  } else if (legendPosition === "bottom" || legendPosition === null) {
+    legendHeight = calculateLegendHeight(
+      panelSchema,
+      chartWidth,
+      chartHeight,
+      seriesData,
+      hasExplicitHeight,
+    );
+    availableHeight = chartHeight - legendHeight;
+  }
+
+  return {
+    availableWidth,
+    availableHeight,
+    legendWidth,
+    legendHeight,
+    hasLegends: true,
+  };
+};
+
+/**
+ * Calculates legend width for right-positioned legends with proper fallbacks
+ * @param {any} panelSchema - Panel configuration
+ * @param {number} chartWidth - Chart container width
+ * @param {number} chartHeight - Chart container height
+ * @param {any[]} seriesData - Series data
+ * @param {boolean} hasExplicitWidth - Whether explicit width is configured
+ * @returns {number} Legend width in pixels
+ */
+const calculateLegendWidth = (
+  panelSchema: any,
+  chartWidth: number,
+  chartHeight: number,
+  seriesData: any[],
+  hasExplicitWidth: boolean,
+): number => {
+  const config = panelSchema.config;
+  const legendsType = config.legends_type;
+  const legendCount = seriesData?.length || 0;
+
+  // Use explicit width if provided
+  if (hasExplicitWidth) {
+    return config.legend_width.unit === "%"
+      ? chartWidth * (config.legend_width.value / 100)
+      : config.legend_width.value;
+  }
+
+  // Handle different legend types
+  if (legendsType === "plain") {
+    return calculateRightLegendWidth(
+      legendCount,
+      chartWidth,
+      chartHeight,
+      seriesData,
+      false,
+    );
+  }
+
+  // Scroll legends - reserve minimum space
+  const minScrollLegendWidth = Math.min(chartWidth * 0.22, 170);
+  return Math.max(minScrollLegendWidth, 120);
+};
+
+/**
+ * Calculates legend height for bottom-positioned legends with proper fallbacks
+ * @param {any} panelSchema - Panel configuration
+ * @param {number} chartWidth - Chart container width
+ * @param {number} chartHeight - Chart container height
+ * @param {any[]} seriesData - Series data
+ * @param {boolean} hasExplicitHeight - Whether explicit height is configured
+ * @returns {number} Legend height in pixels
+ */
+const calculateLegendHeight = (
+  panelSchema: any,
+  chartWidth: number,
+  chartHeight: number,
+  seriesData: any[],
+  hasExplicitHeight: boolean,
+): number => {
+  const config = panelSchema.config;
+  const legendsType = config.legends_type;
+  const legendCount = seriesData?.length || 0;
+
+  // Use explicit height if provided
+  if (hasExplicitHeight) {
+    return config.legend_height.unit === "%"
+      ? chartHeight * (config.legend_height.value / 100)
+      : config.legend_height.value;
+  }
+
+  // Handle different legend types
+  if (legendsType === "plain") {
+    return calculateBottomLegendHeight(
+      legendCount,
+      chartWidth,
+      seriesData,
+      chartHeight,
+    );
+  }
+
+  // Scroll legends - reserve minimum space
+  const minScrollLegendHeight = Math.min(chartHeight * 0.25, 100);
+  return Math.max(minScrollLegendHeight, 60);
+};
+
+/**
+ * Generates CSS grid properties for chart alignment
+ * @param {string} chartAlign - Chart alignment ('left', 'center', or null)
+ * @param {string} legendPosition - Legend position
+ * @param {boolean} shouldApplyAlignment - Whether alignment should be applied
+ * @returns {object} CSS grid properties
+ */
+export const generateChartAlignmentProperties = (
+  chartAlign: string | null,
+  legendPosition: string | null,
+  shouldApplyAlignment: boolean,
+): object => {
+  if (!shouldApplyAlignment || legendPosition !== "right") {
+    return {};
+  }
+
+  const baseProperties = {
+    display: "grid",
+    gridTemplateRows: "1fr",
+  };
+
+  switch (chartAlign) {
+    case "left":
+      return {
+        ...baseProperties,
+        gridTemplateColumns: "minmax(0, 1fr) auto",
+        justifyItems: "start",
+        alignItems: "center",
+        paddingLeft: "5%",
+      };
+
+    case "center":
+      return {
+        ...baseProperties,
+        gridTemplateColumns: "1fr",
+        justifyItems: "center",
+        alignItems: "center",
+      };
+
+    default: // auto - default to center
+      return {
+        ...baseProperties,
+        gridTemplateColumns: "1fr",
+        justifyItems: "center",
+        alignItems: "center",
+      };
+  }
+};
+
+/**
+ * Calculates optimal pie chart radius based on available space and configuration
+ * @param {any} panelSchema - Panel configuration schema
+ * @param {number} availableWidth - Available width for chart
+ * @param {number} availableHeight - Available height for chart
+ * @param {number} originalWidth - Original container width (optional)
+ * @param {number} originalHeight - Original container height (optional)
+ * @returns {number} Optimal radius percentage (0-100)
+ */
+export const calculatePieChartRadius = (
+  panelSchema: any,
+  availableWidth: number,
+  availableHeight: number,
+  originalWidth?: number,
+  originalHeight?: number,
+): number => {
+  // When layout is not available (add panel), calculate based on available dimensions
+  if (!panelSchema.layout) {
+    // Calculate the optimal radius in pixels based on available space
+    const maxAvailableRadius = Math.min(availableWidth, availableHeight) / 2;
+    const optimalRadiusPixels = maxAvailableRadius * 0.75; // 75% of available space for padding
+
+    // If original dimensions are provided, convert to percentage of original container
+    if (originalWidth && originalHeight) {
+      const originalMaxRadius = Math.min(originalWidth, originalHeight) / 2;
+      return Math.min(
+        Math.max((optimalRadiusPixels / originalMaxRadius) * 100, 30),
+        90,
+      );
+    }
+
+    // Fallback: use a conservative percentage
+    return 60;
+  }
+
+  const minRadius = Math.min(
+    panelSchema.layout.w * 30,
+    panelSchema.layout.h * 30,
+  );
+
+  if (minRadius === 0) {
+    return 0;
+  }
+
+  // Calculate radius based on available space
+  const maxRadius = Math.min(availableWidth, availableHeight) / 2;
+  const baseRadius = minRadius / 2;
+
+  // Determine space usage efficiency
+  const isFullSpace =
+    availableWidth === panelSchema.layout.w * 30 &&
+    availableHeight === panelSchema.layout.h * 30;
+  const spaceUsage = isFullSpace ? 0.98 : 0.95;
+
+  const effectiveRadius = Math.min(baseRadius, maxRadius * spaceUsage);
+
+  // Calculate multiplier based on effective radius
+  let multiplier = 100;
+  if (effectiveRadius > 90) multiplier = 120;
+  if (effectiveRadius > 150) multiplier = 140;
+
+  // Calculate final percentage with appropriate limits
+  const maxLimit = isFullSpace ? 98 : 95;
+  return Math.min((effectiveRadius / minRadius) * multiplier, maxLimit);
+};
+
+/**
+ * Determines if chart alignment should be applied based on configuration
+ * @param {any} panelSchema - Panel configuration schema
+ * @param {any[]} seriesData - Series data
+ * @returns {boolean} Whether alignment should be applied
+ */
+export const shouldApplyChartAlignment = (
+  panelSchema: any,
+  seriesData: any[] = [],
+): boolean => {
+  const config = panelSchema.config || {};
+  const showLegends = config.show_legends;
+  const legendPosition = config.legends_position;
+  const legendsType = config.legends_type;
+  const legendCount = seriesData?.length || 0;
+
+  return (
+    showLegends &&
+    legendCount > 0 &&
+    legendPosition === "right" &&
+    (legendsType === "plain" || legendsType === "scroll" || legendsType == null)
+  );
 };
 
 /**
